feat: верстка всех компонентов 2 части проекта (тестовые данные)
