+++ src/components/feed-order-details/FeedOrderDetails.module.css
@@ -13,43 +13,35 @@
 
 .status {
   color: #00cccc;
-  height: 34px;
+  min-height: 34px;
   display: flex;
   align-items: center;
 }
 
-.scrollWrapper {
-  margin: 24px 0 48px;
-  max-height: 366px;
-  overflow-y: scroll;
-  width: 632px;
-  box-sizing: border-box;
-  overflow-x: hidden;
-}
-
-.scrollWrapper::-webkit-scrollbar {
-  width: 8px;
-  background: #2f2f37;
-  height: 192px;
-}
-
-.scrollWrapper::-webkit-scrollbar-thumb {
-  background-color: #8585ad;
-}
-
-.ingredients {
+  .ingredients {
     display: flex;
     flex-direction: column;
-    width: 608px;
     list-style-type: none;
     padding: 0;
-    margin: 4px 0;
+    margin: 24px 0 48px;
     gap: 16px;
-    /* height: 300px; */
+    height: 316px;
+    box-sizing: border-box;
+    overflow-y: scroll;
+    overflow-x: hidden;
+  }
+  
+  .ingredients::-webkit-scrollbar {
+    width: 8px;
+    background: #2f2f37;
+    height: 192px;
+  }
+  
+  .ingredients::-webkit-scrollbar-thumb {
+    background-color: #8585ad;
   }
 .number {
   align-self: center;
-  height: 64px;
   width: 640px;
   display: flex;
   align-items: center;
@@ -61,7 +53,7 @@
   height: 64px;
   object-fit: cover;
   background: linear-gradient(#000000, #000000) padding-box,
-    linear-gradient(63.18deg, #801ab2, #4c4cff) border-box;
+    linear-gradient(70deg, #801ab2, #4c4cff) border-box;
   border-radius: 50%;
   border: 2px solid transparent;
 }
@@ -71,7 +63,10 @@
   align-items: center;
   justify-content: space-between;
   gap: 16px;
-  height: 64px;
+  height: auto;
+  margin-right: 24px;
+  width: 608px
+  
 }
 .name {
   width: 320px;
@@ -80,4 +75,11 @@
 .footer {
     display: flex;
     justify-content: space-between;
+}
+
+.priceContainer {
+  display: flex;
+  gap: 8px;
+  width: 192px;
+  justify-content: end;
 }
