+++ src/components/order-list/OrderList.module.css
@@ -3,7 +3,7 @@
 }
 
 .orderList {
-	max-width: 584px;
+	max-width: calc(100% - 8px);
     width: 100%;
     display: flex;
     flex-direction: column;
@@ -12,6 +12,7 @@
     margin: 0;
     margin-right: 8px;
     gap: 16px;
+    box-sizing: border-box;
     
 }
